Split settings page into focused sub-components (#210, #213)
Settings page (3164 lines) split into 8 focused files:
- settings-shared.tsx: types, constants, and pure helpers
- settings-general-tab.tsx: timezone, logs, density, history reset
- settings-backup-tab.tsx: backup schedule, export, restore
- settings-upgrade-tab.tsx: version check and in-app upgrade
- settings-security-tab.tsx: password change (self-contained)
- settings-logs-tab.tsx: log search and metrics (self-contained)
- settings-support-tab.tsx: support link (self-contained)

Also adds unsaved-changes guards: useBlocker for in-app navigation
and beforeunload for refresh/close when timezone, logs, or backup
schedule fields are dirty. Setup wizard also gets a beforeunload
guard while in pending state.
